requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
gevent==23.9.1
tensorflow==2.15.0
keras==2.15.0
h5py==3.10.0
Pillow==10.0.0
numpy==1.24.3